[List] :danger: changed semantics of "explicit" delete!
up until now, lists were considered to have "delete" set to True by default,
even when it was not explicitly set by the user.
Because of that other parts of the code were behaving as if "!del" was explicitly set -
- in particular, merging: "foo: [1,2] <--- foo: []" was not only clearing the list,
but also removing "foo" altogether, because "foo: whatever <--- foo: !del empty"
is supposed to remove "foo" by design.
Similar bug was also occurring when merging empty dict with an empty dict.

This commit is trying to fix this behaviour.
Specifically, if a resulting child node is empty but "!del" was not set explicitly by the user,
the expected behaviour is to keep the resulting node, otherwise the node should be removed from a dict.

The commit is specified as "Ldanger:" because it is somewhat aggressive in
modifying old, core behaviour,
and as such can potentially break
som edge cases currently not covered by the tests.