[Eval] Regression! Seems like using "locals" with exec break name lookup in certain cases (e.g., list comprehension)
This is because with locals passed any top-level definition is not treated as global!
Therefore visibility of elements defined in the top-level scope
is limited only to this scope - any reference from a sub-scope
tries to refer to a global, but no such global exists!
Very annoying, will probably have to rethink how lazy execution of config
nodes is performed...

Currently thinking about investigating and dynamically rewriting the code to catch all accesses to globals
and then change them to go via a proxy object that will either delegate to normal globals or to the lazily-evaluated config dict obj (current locals)